code/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
httpx==0.28.0
cohere==5.11.4
google-generativeai==0.8.3
qdrant-client==1.12.1
redis==5.2.1
alembic==1.14.0
python-multipart==0.0.17